pyproject.toml
[tool.poetry]
name = "primeboard-backend"
version = "0.1.0"
description = ""
authors = ["Pichugin.Artem <pichugin.artem@wb.ru>"]
readme = "README.md"
#packages = [{include = "primeboard_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.0"
uvicorn = "^0.20.0"
requests = "*"
sqlalchemy = "^1.4.46"
redis = "^4.4.2"
hiredis = "^2.1.1"
schedule = "^1.1.0"
python-dotenv = "^0.21.0"
aiohttp = "^3.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
